+++ python/bind_data_array.h
@@ -1,0 +1,51 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+// Copyright (c) 2021 Scipp contributors (https://github.com/scipp)
+/// @file
+/// @author Simon Heybrock
+#pragma once
+
+#include "scipp/dataset/dataset.h"
+
+#include "pybind11.h"
+
+namespace py = pybind11;
+using namespace scipp;
+
+template <class T, class... Ignored>
+void bind_data_array_properties(py::class_<T, Ignored...> &c) {
+  if constexpr (std::is_same_v<T, DataArray>)
+    c.def_property("name", &T::name, &T::setName,
+                   R"(The name of the held data.)");
+  else
+    c.def_property_readonly("name", &T::name, R"(The name of the held data.)");
+  c.def_property(
+      "data",
+      py::cpp_function([](T &self) { return self.data(); },
+                       py::return_value_policy::move, py::keep_alive<0, 1>()),
+      [](T &self, const VariableConstView &data) { self.data().assign(data); },
+      R"(Underlying data item.)");
+  c.def_property_readonly(
+      "coords",
+      py::cpp_function([](T &self) { return self.coords(); },
+                       py::return_value_policy::move, py::keep_alive<0, 1>()),
+      R"(
+      Dict of aligned coords.)");
+  c.def_property_readonly("meta",
+                          py::cpp_function([](T &self) { return self.meta(); },
+                                           py::return_value_policy::move,
+                                           py::keep_alive<0, 1>()),
+                          R"(
+      Dict of coords and attrs.)");
+  c.def_property_readonly("attrs",
+                          py::cpp_function([](T &self) { return self.attrs(); },
+                                           py::return_value_policy::move,
+                                           py::keep_alive<0, 1>()),
+                          R"(
+      Dict of attrs.)");
+  c.def_property_readonly("masks",
+                          py::cpp_function([](T &self) { return self.masks(); },
+                                           py::return_value_policy::move,
+                                           py::keep_alive<0, 1>()),
+                          R"(
+      Dict of masks.)");
+}
