+++ python/dataset.cpp
@@ -14,6 +14,7 @@
 #include "scipp/dataset/util.h"
 
 #include "bind_data_access.h"
+#include "bind_data_array.h"
 #include "bind_operators.h"
 #include "bind_slice_methods.h"
 #include "detail.h"
@@ -194,42 +195,8 @@ void bind_dataset_view_methods(py::class_<T, Ignored...> &c) {
 }
 
 template <class T, class... Ignored>
-void bind_data_array_properties(py::class_<T, Ignored...> &c) {
-  if constexpr (std::is_same_v<T, DataArray>)
-    c.def_property("name", &T::name, &T::setName,
-                   R"(The name of the held data.)");
-  else
-    c.def_property_readonly("name", &T::name, R"(The name of the held data.)");
-  c.def_property(
-      "data",
-      py::cpp_function([](T &self) { return self.data(); },
-                       py::return_value_policy::move, py::keep_alive<0, 1>()),
-      [](T &self, const VariableConstView &data) { self.data().assign(data); },
-      R"(Underlying data item.)");
-  c.def_property_readonly(
-      "coords",
-      py::cpp_function([](T &self) { return self.coords(); },
-                       py::return_value_policy::move, py::keep_alive<0, 1>()),
-      R"(
-      Dict of aligned coords.)");
-  c.def_property_readonly("meta",
-                          py::cpp_function([](T &self) { return self.meta(); },
-                                           py::return_value_policy::move,
-                                           py::keep_alive<0, 1>()),
-                          R"(
-      Dict of coords and attrs.)");
-  c.def_property_readonly("attrs",
-                          py::cpp_function([](T &self) { return self.attrs(); },
-                                           py::return_value_policy::move,
-                                           py::keep_alive<0, 1>()),
-                          R"(
-      Dict of attrs.)");
-  c.def_property_readonly("masks",
-                          py::cpp_function([](T &self) { return self.masks(); },
-                                           py::return_value_policy::move,
-                                           py::keep_alive<0, 1>()),
-                          R"(
-      Dict of masks.)");
+void bind_data_array(py::class_<T, Ignored...> &c) {
+  bind_data_array_properties(c);
   bind_common_operators(c);
   bind_data_properties(c);
   bind_slice_methods(c);
@@ -333,8 +300,8 @@ void init_dataset(py::module &m) {
                     )");
   options.enable_function_signatures();
 
-  bind_data_array_properties(dataArray);
-  bind_data_array_properties(dataArrayView);
+  bind_data_array(dataArray);
+  bind_data_array(dataArrayView);
 
   py::class_<DatasetConstView>(m, "DatasetConstView")
       .def(py::init<const Dataset &>())
